analysis: remove 't' parameter from StoreM
The 't' parameter in the class `StoreM` was only
used for solving overlapping instances when the
values of the associated type values (e.g. Adr v,
PAdr v, ...) were not known. This was the case
when building a generic monad stack for a
Scheme analysis in `Analysis/Scheme.hs`.

However, this level of parametrisation is actually
not needed as all addresses we use through
the analysis consist of a base (e.g. Ide)
and an optional context (e.g. m-cfa).

Thus, we decided to solve this issue by
making each address unique (as they should)
such that the constraint solver can find
instances for `Adr`, `PAdr`, `VAdr`, ...