analysis: remove 't' from `AllocM`
We argue this point in the same way as the previous commit (cf. 15f2bf5),
the 't' was only there to resolve overlapping instances when composing
the monad stack using general addresses stemming from the abstract
domain (see Adr, PAdr, SAdr, ...).

We solve this problem by making the address types concrete
in the Scheme analysis, thus no longer relying on an instantation
of a type family.