scv: first attempt at defining a symbolic representation for an actor reference.
To support actor references symbolically, the following changes are made:
(1) the symbolic domain gained support for the actor domain.
(2) to this end, a new typeclass is introduced for constraining "(ARef v)"
values (representations of actor pids). This typeclass defines the `identity`
function for these values, which returns a symbolic identity for this function.
(3) the prelude.scm SMTLib script has been updated to include a symbolic
representation of the actor reference in its `V` datatype. A function
called `actor?/v` has been added that provides typing constraints on these
symbolic values.

For (3) more functions need to be added to gain full support for the contract
language. For example, equality constraints between two different pids
need to be added symbolically too.