Accumulated changes
This commit mainly introduces a testing system
for checking the results of a Scheme analysis
against a test specification. For easy of use,
this specification can be defined in the Scheme
file of the SUT directly using a `test-spec`
expression, the `test-spec` expression is
ignored for normal analysis, but used by the
testing harness to validate the output of the
analysis.

The following other changes were also made:
* Fixed `true?` and `false?` which would
erroneously return `Constant true` if a
value would be possibly true, while it could
also be false, for `false?` similar.
* Fixed `choice` implementation in the symbolic
analysis which would run `mcsq` in the false
branch instead of `malt` in the case where
the abstract domain could not precisely
determine the truth value of the condition.
* scv: introduced a `SymbolicStoreT` which
keeps track of symbolic representations
in a local store. This will be needed
for seperating the symbolic representations
from the values in the global store.
* scv: moved stores to a more global position
(i.e., after `runNondetT`) (see above). This also
makes `joinIntra` redundant since the output of
an intra analysis has only one store (a global one,
which is valid over all paths)