feat: meta-programming spawned actors
This commit introduces the ability to intercept message
sends to a particular actor reference by delegating message
sending to the "actor reference" itself.
To this end, we change the translation of actor spawning
to return a lambda instead of an actor reference.
This lambda accepts a message as a parameter and then
sends it to the actual actor reference.

This commit also introduces the following minor changes:
* Suffixed base-level syntax with `^` to emphasize that
it should not be translated anymore.
* Adapted documentation to reflect changes in syntax