fix(simplecc): unsoundness issues leading to undetected contract violations
This commit fixes a number of soundness bugs in the current semantics of
the `SimpleActor` language:

* `EqualLattice` for `UnitLattice` was implement in an unsound way,
it returned `Bottom` even though units are always garantueed to equal
oneanother. This caused issues for the equality of Scheme values that
use these units for representing various values such as the empty list.
* The use of `catch` in `matchList` was incorrect, it caught all errors
  including those that would lead to blame errors or other escape
  effects.

Moreover, the compiler for the SimpleActor language supports lambda's
that contain multiple expressions.