feat(tests): smoke test checking whether the analysis doesn't crash
This commit adds test infrastructure for testing the reliability of
the implemented analyses. More concretely, we implement a smoke
testing infrastructure which checks whether the analysis does not crash
on a predetermined set of inputs. We apply this infrastructure on
analyses for Scheme by checking whether a set of simple Scheme programs
succeed in being analyzed and parsed.