feat(wip): path constraint widening per component
This commit sets the first steps towards widening the path constraint
per component. To this end we apply a similar strategy as the store
widening per component. More specifically, we keep track of the path
constraint not in the component itself but inside a map from components
to path conditions. This map is split into two halves, an "In" part
which denotes the path constraint before entering the component, and an
"Out" part which denotes the path constraint after exiting the
component. The "In" part is computed at the spawn-site of the component,
while the "Out" part is integrated with the state of the component at
the spawn-site.