feat(scv): computation of the model
This commit introduces support for the computation of models from an
SMTLib enabled solver. This has been implemented specifically for the
read-eval-print loop of Z3.

This commit moreover changes the way in which the results from the REPL
of Z3 are parsed. Whereas its output was read until a sentinel string
encountered, the output is parsed as an S-expression datum now.