fix(scv): caching of the model and proper cache keys
This commit adds proper caching for formula's translated using a mapping
of symbolic variables to their abstract counts. To this end we changed
the cache keys to tuples of abstract count mappings and the formula that
needs to be solved.

We also added support for caching the model itself so that it does not
have to be recomputed each time.