fix(ase): various fixes in the small step semantics
This commit introduces various fixes to the small step semantics.

* Proper propagation of the environment in the LetK continuation, where
the extended environment is propagated instead of the original one.
* Correct allocation of addresses in Let2 rule. The previous rule did
not take into account that the context could have changed when applying
the Let2 continuation which could result in an address mismatch.
* Emptying the continuation store so that continuations which should not
be on the continuation stack yet are not resumed when restarting the
computation