fix(analysis): relax functional dependency in StoreM'
Previously the `StoreM'` type class constrained the store contents based
on the monad used. This is too restrictive as the same monad can have
multiple stores in its stack. Hence, the functional dependency is
relaxed so that the monad together with the store type determines the
address and value type.

In theory, this could result in more ambiguities, but in practice
nothing really changed in our implementation. Hence, I think it is safe
to adopt this change into the framework.