feat(scv): symbolic versions of basic predicates
This change is rather big. It requires changing the output of basic
predicates such as `isInteger` to a fixed associated type family member.
To this end the code was also refactored to introduce a single `BoolFor`
type family instead of one for each different core domain (e.g., Boo and
BooS).