fix(simpleactor): various soundness bugs
This commit introduces various soundness fixes:

1. The implementation of `send` erronously only kept the last message
sent as part of the mailbox, causing unsoundness in the contents of the
mailbox.
2. The use of `choice` in `SimpleActor.Semantics` seems to render some
feasible path infeasible, this is currently unresolved and has probably
to do with some symbolic encodings (most notable the encodings of
symbols and their equality), replaced with `cond` for now.