fix(simpleactor): fix soundness bug when applying unknown primitive functions
This commit changes how unknown primitive functions are applied. The
idea is that all primitive functions in the SimpleActor language are
pure (except for the send^ primitive) so that their result can always be
approximated by operations on the input values. This commit joins the
values of the operands into the output result to approximate this fact
and also returns `top`.