feat(simpleactor): translate all contracts to functions
In the previous version of the translation unit, not all contracts were
translated to functions and were instead treated as special syntax when
included in other contracts.

In particular the following issues are solved:
* message contracts in `behavior/c` constructs no longer have to follow
the `message/c` structure. Instead `message/c` structures generates
functions that need to applied with blame labels and message values.
* communication contracts (i.e., ensures/c) in message contracts do no
longer have to be written using an `ensures/c` construct, instead other
language constructs such as variables and function calls can be used.